delete unnecessary hook and change spacing
