+++ app/src/pages/OnDeviceDisplay/ProtocolDetails/index.tsx
@@ -206,6 +206,7 @@ const Summary = (props: {
 }
 
 interface ProtocolSectionContentProps {
+  runId: string
   protocolId: string
   protocolData: ProtocolResource
   currentOption: TabOption
@@ -232,7 +233,7 @@ const ProtocolSectionContent = (
       protocolSection = <Labware protocolId={protocolId} />
       break
     case 'Liquids':
-      protocolSection = <Liquids protocolId={props.protocolId} />
+      protocolSection = <Liquids runId={props.runId} />
       break
     case 'Initial Deck Layout':
       protocolSection = <Deck protocolId={protocolId} />
@@ -307,6 +308,7 @@ export function ProtocolDetails(): JSX.Element | null {
         setCurrentOption={setCurrentOption}
       />
       <ProtocolSectionContent
+        runId={runId}
         protocolId={protocolId}
         protocolData={protocolRecord.data}
         currentOption={currentOption}
